refactor!: remove chain orchestration, adopt coordinator pattern
Drop the entire chain mechanism (chain-io.ts, chain schema, parallel
stages, pause_after, chain_run_id, chain-next proposals) in favor of
the coordinator pattern where the parent agent drives workflows through
individual Agent() calls.

Key changes:
- Delete src/chain-io.ts (784 lines) and test/chain-io.test.ts (1173 lines)
- Strip chain schema and execution from src/index.ts (3088 → 2239 lines)
- Extract 3 reusable guards into src/agent-guards.ts (115 lines)
- Add top-level 'files' param for parallel worker collision detection
- Rewrite all 4 prompt templates for coordinator pattern
- Update README, default-agents, agent-runner, agent-manager

Total src reduction: 8376 → 6858 lines (-18%)
Agent tool params: ~28 nested → 14 flat (-50%)
All 534 tests pass, TypeScript compiles clean.
