+++ src/index.ts
@@ -19,7 +19,7 @@ import { differsFromDefault, diffFromDefault } from "./agent-diff.js";
 import { AgentManager } from "./agent-manager.js";
 import { getAgentConversation, getDefaultExtensions, getDefaultMaxTurns, getGraceTurns, normalizeMaxTurns, setDefaultExtensions, setDefaultMaxTurns, setGraceTurns, steerAgent } from "./agent-runner.js";
 import { BUILTIN_TOOL_NAMES, getAgentConfig, getAllTypes, getAvailableTypes, getDefaultAgentNames, getUserAgentNames, isDefaultsDisabled, registerAgents, resolveType, setDefaultsDisabled } from "./agent-types.js";
-import { buildReadsBlock, type ChainPauseState, consumeChainPauseState, createChainDir, createChainOutputTool, formatChainNextProposal, formatConcurrentActivityNote, formatParallelEditHazardWarning, formatParallelStageEditWarnings, injectOutputInstruction, isAgentReadOnly, isValidChainRunId, mergeParallelOutputs, parseChainNext, persistStepOutput, resolveOutputPath, resolveStepOutput, saveChainPauseState, snapshotOutputFile, substituteChainPlaceholders, validateChainFileOnlyHandoff, validateParallelStage, validateStepIO } from "./chain-io.js";
+import { findOverlappingMemberFiles, formatConcurrentActivityNote, isAgentReadOnly } from "./agent-guards.js";
 import { registerRpcHandlers } from "./cross-extension-rpc.js";
 import { loadCustomAgents } from "./custom-agents.js";
 import { deleteGlobalActivity, setGlobalActivity } from "./global-registry.js";
@@ -212,51 +212,6 @@ function buildNotificationDetails(record: AgentRecord, resultMaxLen: number, act
         : record.result
       : "No output.",
   };
-}
-
-// ---- Chain orchestration helpers ----
-
-type ChainElement =
-  | {
-      subagent_type: string;
-      prompt: string;
-      description?: string;
-      model?: string;
-      thinking?: string;
-      max_turns?: number;
-      output?: string;
-      output_mode?: "inline" | "file-only";
-      reads?: string[];
-      isolated?: boolean;
-      inherit_context?: boolean;
-      isolation?: "worktree";
-      files?: string[];
-      pause_after?: boolean;
-    }
-  | {
-      parallel: Array<{
-        subagent_type: string;
-        prompt: string;
-        description?: string;
-        model?: string;
-        thinking?: string;
-        max_turns?: number;
-        output?: string;
-        output_mode?: "inline" | "file-only";
-        reads?: string[];
-        isolated?: boolean;
-        inherit_context?: boolean;
-        isolation?: "worktree";
-        files?: string[];
-      }>;
-      continue_on_error?: boolean;
-      description?: string;
-      output?: string;
-      output_mode?: "inline" | "file-only";
-    };
-
-function isParallelStage(el: ChainElement): el is Extract<ChainElement, { parallel: unknown[] }> {
-  return Array.isArray((el as { parallel?: unknown[] }).parallel);
 }
 
 export default function (pi: ExtensionAPI) {
@@ -797,8 +752,9 @@ ${buildGuidelinesText()}
 - Use thinking to control extended thinking level.
 - Use inherit_context if the agent needs the parent conversation history.
 - Use isolation: "worktree" to run the agent in an isolated git worktree (safe parallel file modifications).
-- Use chain to run a multi-agent workflow. Each element is either a sequential step (receives the prior output via {previous}) or a {parallel:[...]} stage whose members run concurrently and whose labeled outputs are merged for the next step. Parallel stages fail-fast by default; set continue_on_error to keep surviving outputs.
-- DEFAULT TO PARALLEL. A sequential step blocks until its agent finishes AND pumps that agent's full output into the next step + the context, so back-to-back single-agent steps are the slow, context-bloating path. Only chain steps sequentially when step N+1 literally consumes step N's output. Independent work (multi-file recon, parallel reviews, fan-out edits on non-overlapping files) belongs in ONE {parallel:[...]} stage, never as consecutive single-agent steps. If two adjacent steps don't pass real data between them, they are parallel members, not steps.${scheduleGuideline}`;
+- You are the coordinator. Multi-step work is a loop: dispatch an agent, read its result, synthesize, then write the next agent's prompt. Never paste raw output — extract specific paths, facts, and directives.
+- Parallel work: launch agents in ONE message with run_in_background: true on each. Work is independent when agents edit disjoint files — imports between files are not coupling. Declare files: [...] to detect overlaps. Don't fan out trivial one-line changes.
+- For large handoffs, tell an agent to write its report to a file path (outputFile param), then tell the next agent to read that file.${scheduleGuideline}`;
 
   // Compact Agent tool description (`toolDescriptionMode: "compact"`) — the same
   // load-bearing facts at ~75% fewer tokens, for small/local models. Per-option
@@ -810,8 +766,8 @@ Custom agents: .pi/agents/<name>.md (project) or ${getAgentDir()}/agents/<name>.
 
 Notes:
 - description: 3-5 words (shown in UI). Prompts must be self-contained — the agent has not seen this conversation.
-- Parallel work: one message, multiple Agent calls, run_in_background: true on each. You are notified when background agents finish — never poll or sleep.
-- chain: ordered workflow where each element is a sequential step ({previous} = prior output) or a {parallel:[...]} stage (members run concurrently, outputs merged for the next step). DEFAULT TO PARALLEL: only use a separate sequential step when N+1 consumes N's output. Independent work goes in ONE {parallel:[...]} stage, never consecutive single-agent steps (serial = slow + context bloat).
+- Parallel work: one message, multiple Agent calls, run_in_background: true on each. You are notified when background agents finish — never poll or sleep. Work is independent when agents edit disjoint files; don't fan out trivial changes.
+- Coordinator loop: dispatch agents, read results, synthesize, write next prompt. For handoffs, use the outputFile param so agents can read each other's output.
 - The result is not shown to the user — summarize it for them. Verify an agent's claimed code changes before reporting work done.
 - resume continues a previous agent by ID; steer_subagent messages a running one.
 - isolation: "worktree" runs the agent in an isolated git worktree; changes land on a branch.${scheduleGuideline}`;
@@ -864,134 +820,6 @@ Notes:
     return fullAgentToolDescription;
   })();
 
-  // Shared chain element schema — reused by both `chain` (top-level array) and
-  // When resuming a paused chain, pass remaining steps via the `chain` param.
-  //
-  // The 13 shared step/member fields are extracted into a single object to
-  // minimize source duplication. The TypeBox serializer still inlines them
-  // (no $ref support in pi-ai validation yet), but this keeps descriptions
-  // consistent and makes future $defs dedup a one-line change.
-  const sharedStepMemberFields = {
-    subagent_type: Type.String({
-      description: "Agent type.",
-    }),
-    prompt: Type.String({
-      description: "Task prompt. {previous} = prior step output.",
-    }),
-    description: Type.Optional(
-      Type.String({
-        description: "Short label (shown in UI).",
-      })
-    ),
-    model: Type.Optional(
-      Type.String({
-        description: "Model override.",
-      })
-    ),
-    thinking: Type.Optional(
-      Type.String({
-        description: "Thinking level override.",
-      })
-    ),
-    max_turns: Type.Optional(
-      Type.Number({
-        description: "Max turns.",
-        minimum: 1,
-      })
-    ),
-    output: Type.Optional(
-      Type.String({
-        description: "Output file path. Supports {chain_dir}.",
-      })
-    ),
-    output_mode: Type.Optional(
-      Type.Union(
-        [Type.Literal("inline"), Type.Literal("file-only")],
-        {
-          description: '"inline" (default): pass content. "file-only": write to disk, pass path.',
-        }
-      )
-    ),
-    reads: Type.Optional(
-      Type.Array(
-        Type.String({ description: "File path." }),
-        {
-          description: "Files to prepend as context.",
-        }
-      )
-    ),
-    isolated: Type.Optional(
-      Type.Boolean({
-        description: "No extension/MCP tools.",
-      }),
-    ),
-    inherit_context: Type.Optional(
-      Type.Boolean({
-        description: "Fork parent conversation. Default: false.",
-      }),
-    ),
-    isolation: Type.Optional(
-      Type.Literal("worktree", {
-        description: "Isolated git worktree. Changes saved to branch.",
-      }),
-    ),
-    files: Type.Optional(
-      Type.Array(
-        Type.String({ description: "Owned file path." }),
-        {
-          description: "Declared file ownership. Non-overlapping files suppress clobber warning.",
-        },
-      ),
-    ),
-  } as const;
-
-  const chainElementUnion = Type.Union([
-    Type.Object({
-      ...sharedStepMemberFields,
-      pause_after: Type.Optional(
-        Type.Boolean({
-          description:
-            'Pause chain here. Resume with chain_run_id, pass remaining steps in chain. ' +
-            'A ```chain-next``` block in output becomes a reviewable proposal, never auto-dispatched.',
-        }),
-      ),
-    }),
-    Type.Object({
-      parallel: Type.Array(
-        Type.Object({
-          ...sharedStepMemberFields,
-        }),
-        {
-          description: "Concurrent members. Outputs merged for next step via {previous}.",
-          minItems: 1,
-        }
-      ),
-      continue_on_error: Type.Optional(
-        Type.Boolean({
-          description: "Keep surviving outputs on failure. Default: false (fail-fast).",
-        })
-      ),
-      description: Type.Optional(
-        Type.String({
-          description: "Short label (shown in UI).",
-        })
-      ),
-      output: Type.Optional(
-        Type.String({
-          description: "File path for merged stage output.",
-        })
-      ),
-      output_mode: Type.Optional(
-        Type.Union(
-          [Type.Literal("inline"), Type.Literal("file-only")],
-          {
-            description: '"inline" (default): pass merged content. "file-only": write to disk, pass path.',
-          }
-        )
-      ),
-    }),
-  ]);
-
   pi.registerTool(defineTool({
     name: "Agent",
     label: "Agent",
@@ -1032,19 +860,13 @@ Notes:
           description: "Agent ID to resume.",
         }),
       ),
-      chain: Type.Optional(
-        Type.Array(chainElementUnion, {
-          description:
-            "Multi-agent workflow. Elements: sequential steps ({previous} = prior output) or {parallel:[...]} stages (concurrent, merged). " +
-            "DEFAULT TO PARALLEL: sequential only when N+1 consumes N's output. " +
-            "pause_after stops chain for review; resume with chain_run_id, passing remaining steps here.",
-          minItems: 1,
-        })
-      ),
-      chain_run_id: Type.Optional(
-        Type.String({
-          description: "Resume paused chain by ID. Pass remaining steps in chain.",
-        }),
+      files: Type.Optional(
+        Type.Array(
+          Type.String({ description: "Owned file path." }),
+          {
+            description: "Declared file ownership. Non-overlapping files suppress clobber warning.",
+          },
+        ),
       ),
       isolated: Type.Optional(
         Type.Boolean({
@@ -1067,43 +889,6 @@ Notes:
     // ---- Custom rendering: Claude Code style ----
 
     renderCall(args, theme) {
-      // Chain mode
-      if (args.chain && args.chain.length > 0) {
-        const steps = args.chain as ChainElement[];
-        let text =
-          "▸ " +
-          theme.fg("toolTitle", theme.bold("chain")) +
-          theme.fg("muted", ` (${steps.length} steps)`);
-        for (let i = 0; i < Math.min(steps.length, 4); i++) {
-          const step = steps[i];
-          if (isParallelStage(step)) {
-            text +=
-              `\n  ${theme.fg("muted", `${i + 1}.`)} ` +
-              `${theme.fg("accent", "parallel")} ` +
-              theme.fg(
-                "dim",
-                step.description ?? `${step.parallel.length} members${step.continue_on_error ? ", continue-on-error" : ""}`,
-              );
-            continue;
-          }
-          const name = getDisplayName(
-            resolveType(step.subagent_type) ?? "general-purpose"
-          );
-          const cleanPrompt = step.prompt.replace(/\{previous\}/g, "").replace(/\{chain_dir\}/g, "").trim();
-          const preview =
-            cleanPrompt.length > 45
-              ? `${cleanPrompt.slice(0, 45)}…`
-              : cleanPrompt;
-          text +=
-            `\n  ${theme.fg("muted", `${i + 1}.`)} ` +
-            `${theme.fg("accent", name)} ` +
-            theme.fg("dim", step.description ?? preview);
-        }
-        if (steps.length > 4) {
-          text += `\n  ${theme.fg("muted", `…and ${steps.length - 4} more steps`)}`;
-        }
-        return new Text(text, 0, 0);
-      }
       const displayName = args.subagent_type ? getDisplayName(args.subagent_type) : "Agent";
       const desc = args.description ?? "";
       return new Text("▸ " + theme.fg("toolTitle", theme.bold(displayName)) + (desc ? "  " + theme.fg("muted", desc) : ""), 0, 0);
@@ -1200,31 +985,6 @@ Notes:
 
       // Reload custom agents so new .pi/agents/*.md files are picked up without restart
       reloadCustomAgents();
-
-      if (params.chain_run_id) {
-        if (!params.chain || params.chain.length === 0) {
-          return textResult("chain_run_id requires `chain`: the remaining steps to run after the pause.");
-        }
-        if (!isValidChainRunId(params.chain_run_id)) {
-          return textResult(`Invalid chain_run_id: "${params.chain_run_id}". It must be an ID returned by a previous paused chain, not an arbitrary path.`);
-        }
-        // Single-use: consuming deletes the state file, so a chain_run_id cannot
-        // be resumed twice (which would otherwise re-dispatch the remaining steps
-        // and duplicate whatever file writes/edits they make).
-        const pauseState = consumeChainPauseState(params.chain_run_id);
-        if (!pauseState) {
-          return textResult(`No paused chain found at "${params.chain_run_id}". It may have been cleaned up or already resumed.`);
-        }
-        return await executeChain(params.chain, params, signal, onUpdate, ctx, {
-          chainDir: params.chain_run_id,
-          initialPreviousOutput: pauseState.previousOutput,
-          priorResults: pauseState.results,
-        });
-      }
-
-      if (params.chain && params.chain.length > 0) {
-        return await executeChain(params.chain, params, signal, onUpdate, ctx);
-      }
 
       const rawType = params.subagent_type as SubagentType;
       const resolved = resolveType(rawType);
@@ -1584,615 +1344,6 @@ Notes:
       );
     },
   }));
-
-  async function executeChain(
-    chain: ChainElement[],
-    topParams: any,
-    signal: AbortSignal | undefined,
-    onUpdate: any,
-    ctx: any,
-    resumeState?: { chainDir: string; initialPreviousOutput: string; priorResults: ChainPauseState["results"] }
-  ) {
-    // Validate IO config for all elements before starting any work
-    for (let i = 0; i < chain.length; i++) {
-      const el = chain[i];
-      const ioError = isParallelStage(el)
-        ? validateParallelStage(i, el)
-        : validateStepIO(i, {
-            output: el.output,
-            outputMode: el.output_mode,
-            reads: el.reads,
-          });
-      if (ioError) return textResult(ioError);
-    }
-
-    // Warn (visibly, in chain result) when a file-only step is not matched by reads on the next step
-    const chainWarnings: string[] = validateChainFileOnlyHandoff(chain);
-
-    // Reuse the paused run's scratch directory when resuming, otherwise create a fresh one.
-    const chainDir = resumeState?.chainDir ?? createChainDir();
-
-    let previousOutput = resumeState?.initialPreviousOutput ?? "";
-    const results: Array<{
-      step: number;
-      agent: string;
-      output: string;
-      savedPath?: string;
-      durationMs: number;
-    }> = resumeState ? [...resumeState.priorResults] : [];
-
-    // When resuming a paused chain, step numbers and totals must account for
-    // the steps that already ran before the pause — otherwise resumed steps
-    // restart at "1/N" and collide with the prior run's numbering.
-    const stepBase = resumeState?.priorResults.length ?? 0;
-    const totalSteps = stepBase + chain.length;
-
-    type SequentialChainStep = Extract<ChainElement, { subagent_type: string }>;
-    type ParallelStage = Extract<ChainElement, { parallel: unknown[] }>;
-    type PreparedStep = {
-      resolvedStepType: string;
-      stepDisplayName: string;
-      stepPrompt: string;
-      resolvedOutputPath?: string;
-      outputSnapshot?: ReturnType<typeof snapshotOutputFile>;
-      chainOutputTool?: ReturnType<typeof createChainOutputTool>;
-      readOnly: boolean;
-      resolvedConfig: ReturnType<typeof resolveAgentInvocationConfig>;
-      model: any;
-      effectiveMaxTurns: number | undefined;
-      outputMode: "inline" | "file-only";
-      files?: string[];
-    };
-
-    const prepareStep = (step: SequentialChainStep, priorOutput: string): PreparedStep => {
-      const resolvedStepType = resolveType(step.subagent_type) ?? "general-purpose";
-      const stepDisplayName = getDisplayName(resolvedStepType);
-
-      // Substitute placeholders first, then resolve the output path.
-      const resolvedOutputPath = resolveOutputPath(
-        step.output
-          ? substituteChainPlaceholders(step.output, priorOutput, chainDir)
-          : undefined,
-        ctx.cwd,
-      );
-
-      const outputSnapshot = snapshotOutputFile(resolvedOutputPath);
-
-      const resolvedReads = step.reads?.map((p) => substituteChainPlaceholders(p, priorOutput, chainDir));
-      const readsBlock = buildReadsBlock(resolvedReads, ctx.cwd);
-
-      let stepPrompt = substituteChainPlaceholders(step.prompt, priorOutput, chainDir);
-      if (readsBlock) {
-        stepPrompt = readsBlock + stepPrompt;
-      }
-      stepPrompt = injectOutputInstruction(stepPrompt, resolvedOutputPath);
-
-      // Cross-dispatch awareness: chain-internal parallel validation only sees
-      // members of THIS chain's own stage. A sibling top-level Agent call (a
-      // separate, concurrently running chain sharing the same cwd) is invisible
-      // to it. If one is active and writable and not worktree-isolated, tell
-      // this step so it doesn't misattribute the sibling's legitimate changes.
-      const concurrentWriters = manager.listAgents().filter((a) => {
-        if (a.status !== "running" || a.worktree) return false;
-        const otherConfig = getAgentConfig(a.type);
-        return !isAgentReadOnly(otherConfig?.builtinToolNames, otherConfig?.disallowedTools);
-      }).length;
-      if (concurrentWriters > 0) {
-        stepPrompt = `${stepPrompt}\n\n${formatConcurrentActivityNote(concurrentWriters)}`;
-      }
-
-      const customConfig = getAgentConfig(resolvedStepType);
-      const chainOutputTool = resolvedOutputPath ? createChainOutputTool(resolvedOutputPath) : undefined;
-      const readOnly = isAgentReadOnly(customConfig?.builtinToolNames, customConfig?.disallowedTools);
-      const stepParams = {
-        subagent_type: resolvedStepType,
-        description: step.description,
-        model: step.model ?? topParams.model,
-        thinking: step.thinking ?? topParams.thinking,
-        max_turns: step.max_turns ?? topParams.max_turns,
-        isolated: step.isolated ?? topParams.isolated,
-        inherit_context: step.inherit_context ?? topParams.inherit_context,
-        isolation: step.isolation ?? topParams.isolation,
-      };
-      const resolvedConfig = resolveAgentInvocationConfig(customConfig, stepParams);
-
-      let model = ctx.model;
-      if (resolvedConfig.modelInput) {
-        const resolved = resolveModel(resolvedConfig.modelInput, ctx.modelRegistry);
-        if (typeof resolved !== "string") {
-          model = resolved;
-        }
-      }
-
-      const effectiveMaxTurns = normalizeMaxTurns(resolvedConfig.maxTurns ?? getDefaultMaxTurns());
-      const outputMode = step.output_mode ?? "inline";
-
-      // Inject turn budget so chain steps can pace themselves.
-      if (effectiveMaxTurns != null) {
-        stepPrompt += `\n\n[Turn budget: ${effectiveMaxTurns} turns. Pace yourself \u2014 land the core change first, then validate if turns remain.]`;
-      }
-
-      return {
-        resolvedStepType,
-        stepDisplayName,
-        stepPrompt,
-        resolvedOutputPath,
-        outputSnapshot,
-        chainOutputTool,
-        readOnly,
-        resolvedConfig,
-        model,
-        effectiveMaxTurns,
-        outputMode,
-        // Normalize declared file ownership to absolute paths so overlap
-        // detection compares like-for-like (a member declaring "src/a.ts"
-        // and another declaring the absolute equivalent must still collide).
-        files: step.files?.map((f) => resolveOutputPath(f, ctx.cwd)).filter((f): f is string => !!f),
-      };
-    };
-
-    const runParallelStage = async (
-      stage: ParallelStage,
-      stageIndex: number,
-      stageInput: string,
-    ): Promise<
-      | { fatal: true; fatalMessage: string }
-      | { fatal: false; previousOutput: string; results: Array<{ step: number; agent: string; output: string; savedPath?: string; durationMs: number }>; warnings: string[] }
-    > => {
-      if (signal?.aborted) {
-        return { fatal: true, fatalMessage: `Chain stopped at step ${stepBase + stageIndex + 1}: aborted.` };
-      }
-
-      const warnings: string[] = [];
-      const preparedMembers = stage.parallel.map((member) => prepareStep(member as SequentialChainStep, stageInput));
-
-      const seenOutputPaths = new Map<string, number>();
-      for (let m = 0; m < preparedMembers.length; m++) {
-        const prepared = preparedMembers[m];
-        if (prepared.resolvedOutputPath) {
-          const prev = seenOutputPaths.get(prepared.resolvedOutputPath);
-          if (prev !== undefined) {
-            warnings.push(
-              `⚠ Parallel stage ${stepBase + stageIndex + 1} members ${prev + 1} and ${m + 1} both write ${prepared.resolvedOutputPath}; last write wins.`,
-            );
-          } else {
-            seenOutputPaths.set(prepared.resolvedOutputPath, m);
-          }
-        }
-      }
-
-      warnings.push(
-        ...formatParallelStageEditWarnings(
-          stageIndex,
-          preparedMembers.map((prepared, idx) => ({
-            memberIndex: idx,
-            displayName: prepared.stepDisplayName,
-            readOnly: prepared.readOnly,
-            isWorktree: prepared.resolvedConfig.isolation === "worktree",
-            files: prepared.files,
-          })),
-        ),
-      );
-
-      const memberStates: Array<AgentActivity | undefined> = new Array(preparedMembers.length);
-      const memberAgentIds: Array<string | undefined> = new Array(preparedMembers.length);
-      const spawnedAgentIds: Array<string | undefined> = new Array(preparedMembers.length);
-      const memberStartedAt: Array<number | undefined> = new Array(preparedMembers.length);
-      const records: Array<AgentRecord | undefined> = new Array(preparedMembers.length);
-
-      const spinnerInterval = setInterval(() => {
-        onUpdate?.({
-          content: [
-            {
-              type: "text",
-              text:
-                `Chain step ${stepBase + stageIndex + 1}/${totalSteps} — parallel: ` +
-                preparedMembers
-                  .map((prepared, idx) => {
-                    const state = memberStates[idx];
-                    return `${prepared.stepDisplayName}: ${state ? describeActivity(state.activeTools, state.responseText) : "starting…"}`;
-                  })
-                  .join("  |  "),
-            },
-          ],
-        });
-      }, 80);
-
-      const finalizeParallelMembers = () => {
-        clearInterval(spinnerInterval);
-        const finishedIds = new Set<string>();
-        for (const id of memberAgentIds) {
-          if (id) finishedIds.add(id);
-        }
-        for (const id of spawnedAgentIds) {
-          if (id) finishedIds.add(id);
-        }
-        for (const id of finishedIds) {
-          agentActivity.delete(id);
-          deleteGlobalActivity(id);
-          widget.markFinished(id);
-        }
-        widget.update();
-      };
-
-      try {
-        for (let m = 0; m < preparedMembers.length; m++) {
-          const prepared = preparedMembers[m];
-          const { state, callbacks } = createActivityTracker(prepared.effectiveMaxTurns, () => {
-            onUpdate?.({
-              content: [
-                {
-                  type: "text",
-                  text: `Chain step ${stepBase + stageIndex + 1}/${totalSteps} parallel ${m + 1}/${preparedMembers.length}: ${prepared.stepDisplayName}: ${describeActivity(state.activeTools, state.responseText)}`,
-                },
-              ],
-            });
-          });
-          memberStates[m] = state;
-
-          const origOnSession = callbacks.onSessionCreated;
-          callbacks.onSessionCreated = (session: any) => {
-            origOnSession(session);
-            for (const a of manager.listAgents()) {
-              if (a.session === session) {
-                memberAgentIds[m] = a.id;
-                agentActivity.set(a.id, state);
-                setGlobalActivity(a.id, state);
-                widget.ensureTimer();
-                widget.update();
-                break;
-              }
-            }
-          };
-
-          memberStartedAt[m] = Date.now();
-          const member = stage.parallel[m];
-          const id = manager.spawn(pi, ctx, prepared.resolvedStepType, prepared.stepPrompt, {
-            description: member.description ?? `Parallel member ${m + 1}: ${prepared.stepDisplayName}`,
-            model: prepared.model,
-            maxTurns: prepared.effectiveMaxTurns,
-            isolated: prepared.resolvedConfig.isolated,
-            inheritContext: prepared.resolvedConfig.inheritContext,
-            thinkingLevel: prepared.resolvedConfig.thinking,
-            isolation: prepared.resolvedConfig.isolation,
-            signal,
-            customTools: prepared.chainOutputTool ? [prepared.chainOutputTool] : undefined,
-            isBackground: false,
-            ...callbacks,
-          });
-          spawnedAgentIds[m] = id;
-          const record = manager.getRecord(id);
-          if (!record) {
-            throw new Error(`Parallel stage ${stepBase + stageIndex + 1}: failed to capture spawned member ${m + 1}.`);
-          }
-          records[m] = record;
-        }
-      } catch (err) {
-        const abortedIds = new Set<string>();
-        for (const id of memberAgentIds) {
-          if (id) abortedIds.add(id);
-        }
-        for (const id of spawnedAgentIds) {
-          if (id) abortedIds.add(id);
-        }
-        for (const id of abortedIds) {
-          manager.abort(id);
-        }
-        finalizeParallelMembers();
-        return {
-          fatal: true,
-          fatalMessage: `Chain failed at step ${stepBase + stageIndex + 1}/${totalSteps} parallel stage start: ${err instanceof Error ? err.message : String(err)}`,
-        };
-      }
-
-      await Promise.allSettled(records.map((r) => r!.promise!));
-
-      const memberRows: Array<{
-        step: number;
-        agent: string;
-        summaryAgent: string;
-        output: string;
-        mergeOutput: string;
-        savedPath?: string;
-        durationMs: number;
-        error?: string;
-      }> = [];
-      let fatalMessage: string | undefined;
-
-      for (let m = 0; m < preparedMembers.length; m++) {
-        const prepared = preparedMembers[m];
-        const record = records[m]!;
-        const durationMs = (record.completedAt ?? Date.now()) - (memberStartedAt[m] ?? Date.now());
-        const rawOutput = record.result?.trim() ?? "";
-
-        if (record.status === "error" || record.status === "stopped" || record.status === "aborted") {
-          const errMsg = record.error ?? record.status;
-          if (!stage.continue_on_error) {
-            fatalMessage ??= `Chain failed at step ${stepBase + stageIndex + 1}/${totalSteps} parallel member ${m + 1} (${prepared.stepDisplayName}): ${errMsg}`;
-          } else {
-            warnings.push(`⚠ Stage ${stepBase + stageIndex + 1} member ${m + 1} (${prepared.stepDisplayName}) failed: ${errMsg}. Surviving outputs preserved.`);
-          }
-          memberRows.push({
-            step: stepBase + stageIndex + 1,
-            agent: prepared.stepDisplayName,
-            summaryAgent: `${prepared.stepDisplayName} (parallel ${m + 1}/${preparedMembers.length})`,
-            output: rawOutput || "(no output)",
-            mergeOutput: rawOutput,
-            durationMs,
-            error: errMsg,
-          });
-          continue;
-        }
-
-        const { output: resolvedOutput, savedPath, saveError } = resolveStepOutput({
-          outputPath: prepared.resolvedOutputPath,
-          fallbackOutput: rawOutput,
-          snapshot: prepared.outputSnapshot,
-        });
-
-        if (saveError) {
-          console.warn(`[pi-subagents] Parallel stage ${stepBase + stageIndex + 1} member ${m + 1} output file warning: ${saveError}`);
-          const fallbackMsg = prepared.readOnly
-            ? `ℹ Parallel stage ${stepBase + stageIndex + 1} member ${m + 1} used a read-only agent; output was saved by the orchestrator to ${savedPath ?? prepared.resolvedOutputPath}.`
-            : `ℹ Parallel stage ${stepBase + stageIndex + 1} member ${m + 1}: agent did not write to the output file; in-memory output was saved by the orchestrator to ${savedPath ?? prepared.resolvedOutputPath}.`;
-          warnings.push(fallbackMsg);
-        }
-
-        memberRows.push({
-          step: stepBase + stageIndex + 1,
-          agent: prepared.stepDisplayName,
-          summaryAgent: `${prepared.stepDisplayName} (parallel ${m + 1}/${preparedMembers.length})`,
-          output: prepared.outputMode === "file-only" && savedPath ? `Output saved to: ${savedPath}` : resolvedOutput,
-          mergeOutput: resolvedOutput,
-          savedPath,
-          durationMs,
-        });
-      }
-
-      if (fatalMessage) {
-        const completed = memberRows.filter((row) => !row.error);
-        const completedSummary = completed.length > 0
-          ? `Completed members:\n${completed.map((row) => `  ${row.step}. ${row.summaryAgent}: ${row.output.slice(0, 100)}…`).join("\n")}`
-          : "";
-        const priorSummary = results.length > 0
-          ? `Completed steps:\n${results.map((r) => `  ${r.step}. ${r.agent}: ${r.output.slice(0, 100)}…`).join("\n")}`
-          : "";
-        finalizeParallelMembers();
-        return {
-          fatal: true,
-          fatalMessage:
-            fatalMessage +
-            (completedSummary ? `\n\n${completedSummary}` : "") +
-            (priorSummary ? `\n\n${priorSummary}` : ""),
-        };
-      }
-
-      const merged = mergeParallelOutputs(memberRows.map((row) => ({
-        agent: row.agent,
-        output: row.mergeOutput,
-        error: row.error,
-      })));
-
-      const resolvedStageOutputPath = resolveOutputPath(
-        stage.output ? substituteChainPlaceholders(stage.output, stageInput, chainDir) : undefined,
-        ctx.cwd,
-      );
-      let stageSavedPath: string | undefined;
-      if (resolvedStageOutputPath) {
-        const persist = persistStepOutput(resolvedStageOutputPath, merged);
-        stageSavedPath = persist.savedPath;
-        if (persist.saveError) {
-          warnings.push(`⚠ Parallel stage ${stepBase + stageIndex + 1} merged output save error: ${persist.saveError}`);
-        }
-      }
-
-      const previousForNext = stage.output_mode === "file-only" && stageSavedPath ? stageSavedPath : merged;
-      const rows = memberRows.map((row) => ({
-        step: row.step,
-        agent: row.summaryAgent,
-        output: row.output,
-        savedPath: row.savedPath,
-        durationMs: row.durationMs,
-      }));
-
-      finalizeParallelMembers();
-      return {
-        fatal: false,
-        previousOutput: previousForNext,
-        results: rows,
-        warnings,
-      };
-    };
-
-    for (let i = 0; i < chain.length; i++) {
-      const el = chain[i];
-      if (signal?.aborted) {
-        return textResult(`Chain stopped at step ${stepBase + i + 1}: aborted.`);
-      }
-
-      if (isParallelStage(el)) {
-        const stageResult = await runParallelStage(el, i, previousOutput);
-        if (stageResult.fatal) return textResult(stageResult.fatalMessage);
-        previousOutput = stageResult.previousOutput;
-        results.push(...stageResult.results);
-        chainWarnings.push(...stageResult.warnings);
-        continue;
-      }
-
-      const prepared = prepareStep(el, previousOutput);
-
-      const { state: stepState, callbacks: stepCallbacks } =
-        createActivityTracker(prepared.effectiveMaxTurns, () => {
-          onUpdate?.({
-            content: [
-              {
-                type: "text",
-                text: `Chain step ${stepBase + i + 1}/${totalSteps}: ${stepState.toolUses} tool uses…`,
-              },
-            ],
-          });
-        });
-
-      let stepAgentId: string | undefined;
-      const origStepOnSession = stepCallbacks.onSessionCreated;
-      stepCallbacks.onSessionCreated = (session: any) => {
-        origStepOnSession(session);
-        for (const a of manager.listAgents()) {
-          if (a.session === session) {
-            stepAgentId = a.id;
-            agentActivity.set(a.id, stepState);
-            setGlobalActivity(a.id, stepState);
-            widget.ensureTimer();
-            widget.update();
-            break;
-          }
-        }
-      };
-
-      const stepStartedAt = Date.now();
-      const spinnerInterval = setInterval(() => {
-        onUpdate?.({
-          content: [
-            {
-              type: "text",
-              text: `Chain step ${stepBase + i + 1}/${totalSteps} — ${prepared.stepDisplayName}: ${describeActivity(stepState.activeTools, stepState.responseText)}`,
-            },
-          ],
-        });
-      }, 80);
-
-      let record: AgentRecord;
-      try {
-        record = await manager.spawnAndWait(
-          pi,
-          ctx,
-          prepared.resolvedStepType,
-          prepared.stepPrompt,
-          {
-            description: el.description ?? `Chain step ${stepBase + i + 1}: ${prepared.stepDisplayName}`,
-            model: prepared.model,
-            maxTurns: prepared.effectiveMaxTurns,
-            isolated: prepared.resolvedConfig.isolated,
-            inheritContext: prepared.resolvedConfig.inheritContext,
-            thinkingLevel: prepared.resolvedConfig.thinking,
-            isolation: prepared.resolvedConfig.isolation,
-            signal,
-            customTools: prepared.chainOutputTool ? [prepared.chainOutputTool] : undefined,
-            ...stepCallbacks,
-          }
-        );
-      } finally {
-        clearInterval(spinnerInterval);
-        if (stepAgentId) {
-          agentActivity.delete(stepAgentId);
-          deleteGlobalActivity(stepAgentId);
-          widget.markFinished(stepAgentId);
-          widget.update();
-        }
-      }
-
-      const stepDurationMs = (record.completedAt ?? Date.now()) - stepStartedAt;
-
-      if (record.status === "error") {
-        return textResult(
-          `Chain failed at step ${stepBase + i + 1}/${totalSteps} (${prepared.stepDisplayName}): ${record.error}\n\n` +
-            (results.length > 0
-              ? `Completed steps:\n${results.map((r) => `  ${r.step}. ${r.agent}: ${r.output.slice(0, 100)}…`).join("\n")}`
-              : "")
-        );
-      }
-
-      if (record.status === "stopped") {
-        return textResult(
-          `Chain stopped at step ${stepBase + i + 1}/${totalSteps} (${prepared.stepDisplayName}): agent was stopped.`
-        );
-      }
-
-      if (record.status === "aborted") {
-        return textResult(`Chain aborted at step ${stepBase + i + 1}/${totalSteps}.`);
-      }
-
-      const inMemoryOutput = record.result?.trim() ?? "";
-      const { output: resolvedOutput, savedPath, saveError } = resolveStepOutput({
-        outputPath: prepared.resolvedOutputPath,
-        fallbackOutput: inMemoryOutput,
-        snapshot: prepared.outputSnapshot,
-      });
-
-      if (prepared.outputMode === "file-only" && savedPath) {
-        previousOutput = savedPath;
-      } else {
-        previousOutput = resolvedOutput;
-      }
-
-      results.push({
-        step: stepBase + i + 1,
-        agent: prepared.stepDisplayName,
-        output: prepared.outputMode === "file-only" && savedPath
-          ? `Output saved to: ${savedPath}`
-          : resolvedOutput,
-        savedPath,
-        durationMs: stepDurationMs,
-      });
-
-      if (saveError) {
-        console.warn(`[pi-subagents] Chain step ${stepBase + i + 1} output file warning: ${saveError}`);
-        const fallbackMsg = prepared.readOnly
-          ? `ℹ Step ${stepBase + i + 1} used a read-only agent; output was saved by the orchestrator to ${savedPath ?? prepared.resolvedOutputPath}.`
-          : `ℹ Step ${stepBase + i + 1}: agent did not write to the output file; in-memory output was saved by the orchestrator to ${savedPath ?? prepared.resolvedOutputPath}.`;
-        chainWarnings.push(fallbackMsg);
-      }
-
-      if (el.pause_after) {
-        saveChainPauseState(chainDir, { previousOutput, results, pausedAtStep: i });
-        const pausedMs = results.reduce((s, r) => s + r.durationMs, 0);
-        const pausedSummary = results
-          .map((r) => {
-            const fileNote = r.savedPath ? `\n  → saved to ${r.savedPath}` : "";
-            return `Step ${r.step} (${r.agent}) — ${formatMs(r.durationMs)}:${fileNote}\n${r.output.slice(0, 200)}${r.output.length > 200 ? "…" : ""}`;
-          })
-          .join("\n\n---\n\n");
-
-        const chainNext = parseChainNext(previousOutput);
-        let proposalBlock: string;
-        if (chainNext.errors.length > 0) {
-          proposalBlock =
-            `\n\n⚠ chain-next block invalid, no auto-proposal:\n${chainNext.errors.map((e) => `  - ${e}`).join("\n")}\n\n` +
-            `Review ${prepared.stepDisplayName}'s output above and author the remaining steps yourself, then call Agent again with chain_run_id: "${chainDir}" and chain: [...steps].`;
-        } else if (chainNext.proposal) {
-          proposalBlock =
-            `\n\n**Proposed next steps** (from ${prepared.stepDisplayName}'s chain-next block — review before dispatching, do not paste through unreviewed):\n${formatChainNextProposal(chainNext.proposal)}\n\n` +
-            `To continue: call Agent again with chain_run_id: "${chainDir}" and chain: [...] built from (or overriding) this proposal — e.g. one { parallel: [...] } stage using these entries.`;
-        } else {
-          proposalBlock =
-            `\n\nNo chain-next block found in this step's output. Review the output above and author the remaining steps yourself, then call Agent again with chain_run_id: "${chainDir}" and chain: [...steps].`;
-        }
-
-        return textResult(
-          `Chain paused after step ${stepBase + i + 1}/${totalSteps} (${prepared.stepDisplayName}). chain_run_id: ${chainDir}\n\n${pausedSummary}${proposalBlock}`
-        );
-      }
-    }
-
-    const totalMs = results.reduce((s, r) => s + r.durationMs, 0);
-    const summary = results
-      .map(
-        (r) => {
-          const fileNote = r.savedPath ? `\n  → saved to ${r.savedPath}` : "";
-          return `Step ${r.step} (${r.agent}) — ${formatMs(r.durationMs)}:${fileNote}\n${r.output.slice(0, 200)}${r.output.length > 200 ? "…" : ""}`;
-        }
-      )
-      .join("\n\n---\n\n");
-
-    const warningsBlock =
-      chainWarnings.length > 0
-        ? `\n\n**Warnings:**\n${chainWarnings.map((w) => `  ${w}`).join("\n")}`
-        : "";
-
-    return textResult(
-      `Chain completed in ${formatMs(totalMs)} (${totalSteps} steps). Chain directory: ${chainDir}\n\n${summary}${warningsBlock}`
-    );
-  }
 
   // ---- get_subagent_result tool ----
